+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Added: `selector-attribute-operator-space-before` rule.
 - Added: `selector-max-empty-lines` rule.
 - Added: `selector-pseudo-element-no-unknown` rule.
+- Added: `selector-max-compound-selectors` rule.
 - Added: flexible support for end-of-line comments in `at-rule-semicolon-newline-after`, `block-opening-brace-newline-after`, and `declaration-block-semicolon-newline-after`.
 - Fixed: string and verbose formatters no longer use an ambiguous colour schemes.
 - Fixed: string formatter no longer outputs an empty line if there are no problems.
@@ -54,7 +55,6 @@
 - Fixed: `number-max-precision` now ignores `@import` at-rules and `url()` functions.
 - Fixed: `selector-class-pattern` and `selector-id-pattern` rules now ignore SCSS variable interpolation.
 - Fixed: `value-list-comma-*` rules now ignore SCSS maps.
-- Added: `selector-max-depth` rule.
 
 # 6.2.2